Move store.go to internal/store
