+++ plugin/store.go
@@ -1,13 +1,6 @@
-package main
+package plugin
 
-import (
-	"context"
-	"errors"
-)
-
-var (
-	ErrEndpointDataNotFound = errors.New("endpoint data not found")
-)
+import "context"
 
 type Store interface {
 	Get(ctx context.Context, key string) (string, error)
